backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.9.0
python-multipart==0.0.20
sendgrid==6.11.0
httpx==0.27.0
apscheduler==3.10.4
ruff==0.4.4
black==26.3.1
pytest==8.2.0
pytest-asyncio==0.23.6
reportlab==4.2.5
openpyxl==3.1.5
aiofiles==23.2.1
weasyprint==65.1
pydyf==0.11.0
jinja2==3.1.6
